--- src/main/java/graduate/schedule/dto/web/request/ChangeScheduleRequestDTO.java
@@ -1,11 +1,0 @@
-package graduate.schedule.dto.web.request;
-
-import lombok.Getter;
-
-@Getter
-public class ChangeScheduleRequestDTO {
-    private Long scheduleId;
-    private Long employeeId;
-    private String startTime; //HH:mm
-    private String endTime; //HH:mm
-}
