+++ src/main/java/graduate/schedule/dto/web/request/executive/CreateScheduleRequestDTO.java
@@ -1,0 +1,14 @@
+package graduate.schedule.dto.web.request.executive;
+
+import lombok.Getter;
+
+import java.sql.Date;
+
+@Getter
+public class CreateScheduleRequestDTO {
+    private Long storeId;
+    private Date date;
+    private Long employeeId;
+    private String startTime; //HH:mm
+    private String endTime; //HH:mm
+}
