+++ src/main/java/graduate/schedule/dto/web/request/executive/SetMemberGradeRequestDTO.java
@@ -1,4 +1,4 @@
-package graduate.schedule.dto.web.request;
+package graduate.schedule.dto.web.request.executive;
 
 import lombok.Getter;
