+++ src/main/java/graduate/schedule/service/ExecutiveService.java
@@ -7,11 +7,11 @@ import graduate.schedule.common.exception.StoreScheduleException;
 import graduate.schedule.domain.member.Member;
 import graduate.schedule.domain.store.*;
 import graduate.schedule.dto.store.EmployeeDTO;
-import graduate.schedule.dto.web.request.ChangeScheduleRequestDTO;
-import graduate.schedule.dto.web.request.DeleteStoreMemberRequestDTO;
-import graduate.schedule.dto.web.request.SetMemberGradeRequestDTO;
-import graduate.schedule.dto.web.response.ChangeScheduleResponseDTO;
-import graduate.schedule.dto.web.response.StoreAllEmployeeResponseDTO;
+import graduate.schedule.dto.web.request.executive.ChangeScheduleRequestDTO;
+import graduate.schedule.dto.web.request.executive.DeleteStoreMemberRequestDTO;
+import graduate.schedule.dto.web.request.executive.SetMemberGradeRequestDTO;
+import graduate.schedule.dto.web.response.executive.ChangeScheduleResponseDTO;
+import graduate.schedule.dto.web.response.executive.StoreAllEmployeeResponseDTO;
 import graduate.schedule.repository.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
